2.0.0.2:  - video_format: Refactored and added duration and time_scale.  - decklink_consumer: Refactored.  - decklink_prodicer: Refactored.  - amcp: Added log message to PLAY command.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@714 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: d100d17e81ed61c7638b681989cfdab5276430db

+++ common/diagnostics/graph.h
@@ -56,11 +56,11 @@ class graph
 	graph(const std::string& name);
 	graph(const printer& parent_printer);
 public:
-	void update_value(const std::string& name, float value);
-	void set_value(const std::string& name, float value);
+	void update_value(const std::string& name, double value);
+	void set_value(const std::string& name, double value);
 	void set_color(const std::string& name, color c);
 	void add_tag(const std::string& name);
-	void add_guide(const std::string& name, float value);
+	void add_guide(const std::string& name, double value);
 private:
 	struct implementation;
 	std::shared_ptr<implementation> impl_;
